: Address review: point docs at the target state, move execution-path timeout constants to timeouts.py
